Struct block_cache::BlockCache

pub struct BlockCache { /* fields omitted */ }

A cache to store read and written blocks from a storage device.

Implementations

impl BlockCache

pub fn new(storage_device: StorageDeviceRef) -> BlockCache

Creates a new BlockCache device

pub fn flush(&mut self, block_num: Option<usize>) -> Result<(), &'static str>

Flushes the given block to the backing storage device. If the block_to_flush is None, all blocks in the entire cache will be written back to the storage device.

pub fn read_block<'c>(
cache: &'c mut BlockCache,
block: usize
) -> Result<&'c [u8], &'static str>

An internal function that first checks the cache for a specific block in order to avoid reading from the storage device. If that block exists in the cache, it is copied into the buffer. If not, it is read from the storage device into the cache, and then copied into the buffer.

pub fn write_block(
&mut self,
block_num: usize,
buffer_to_write: Cow<'_, [u8]>
) -> Result<(), &'static str>